pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "fitz>=0.0.1.dev2",
    "openai>=1.79.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-pptx>=1.0.2",
    "streamlit>=1.45.1",
]
